refactor: refactor to HeadRequestHandler
